requirements.txt
python-telegram-bot==20.*
wikipedia-api
psycopg2-binary
python-dotenv
